replace Montgomery Goldilocks with native implementation
