+++ crypto/math/benches/goldilocks_benchmark.rs
@@ -1,32 +1,17 @@
-//! Benchmark comparing Goldilocks Montgomery vs Native implementations
-//!
-//! Validates that Native form performs better than Montgomery form for field operations.
+//! Benchmark for native Goldilocks field operations.
 
 use std::hint::black_box;
 
 use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
-use rand::RngCore;
-use rand::SeedableRng;
-
-// Montgomery form (base field)
-use math::field::element::FieldElement;
-use math::field::fields::fft_friendly::u64_goldilocks::U64GoldilocksPrimeField;
+use rand::{RngCore, SeedableRng};
 
 // Native form (base field)
-use math::field::fields::fft_friendly::u64_goldilocks_native::GoldilocksField;
+use math::field::element::FieldElement;
+use math::field::fields::fft_friendly::u64_goldilocks::GoldilocksField;
 
-type MontFE = FieldElement<U64GoldilocksPrimeField>;
 type NativeFE = FieldElement<GoldilocksField>;
 
 const SIZES: [usize; 5] = [100, 1000, 10000, 100000, 1000000];
-
-/// Generate random Montgomery field element pairs with fixed seed for reproducibility
-fn rand_mont_elements(num: usize) -> Vec<(MontFE, MontFE)> {
-    let mut rng = rand_chacha::ChaCha20Rng::seed_from_u64(9001);
-    (0..num)
-        .map(|_| (MontFE::from(rng.next_u64()), MontFE::from(rng.next_u64())))
-        .collect()
-}
 
 /// Generate random Native field element pairs with fixed seed for reproducibility
 fn rand_native_elements(num: usize) -> Vec<(NativeFE, NativeFE)> {
@@ -45,20 +30,7 @@ fn bench_add(c: &mut Criterion) {
     let mut group = c.benchmark_group("goldilocks_add");
 
     for size in SIZES {
-        let mont_data = rand_mont_elements(size);
         let native_data = rand_native_elements(size);
-
-        group.bench_with_input(
-            BenchmarkId::new("montgomery", size),
-            &mont_data,
-            |b, data| {
-                b.iter(|| {
-                    for (x, y) in data {
-                        black_box(black_box(x) + black_box(y));
-                    }
-                })
-            },
-        );
 
         group.bench_with_input(BenchmarkId::new("native", size), &native_data, |b, data| {
             b.iter(|| {
@@ -75,20 +47,7 @@ fn bench_sub(c: &mut Criterion) {
     let mut group = c.benchmark_group("goldilocks_sub");
 
     for size in SIZES {
-        let mont_data = rand_mont_elements(size);
         let native_data = rand_native_elements(size);
-
-        group.bench_with_input(
-            BenchmarkId::new("montgomery", size),
-            &mont_data,
-            |b, data| {
-                b.iter(|| {
-                    for (x, y) in data {
-                        black_box(black_box(x) - black_box(y));
-                    }
-                })
-            },
-        );
 
         group.bench_with_input(BenchmarkId::new("native", size), &native_data, |b, data| {
             b.iter(|| {
@@ -105,20 +64,7 @@ fn bench_mul(c: &mut Criterion) {
     let mut group = c.benchmark_group("goldilocks_mul");
 
     for size in SIZES {
-        let mont_data = rand_mont_elements(size);
         let native_data = rand_native_elements(size);
-
-        group.bench_with_input(
-            BenchmarkId::new("montgomery", size),
-            &mont_data,
-            |b, data| {
-                b.iter(|| {
-                    for (x, y) in data {
-                        black_box(black_box(x) * black_box(y));
-                    }
-                })
-            },
-        );
 
         group.bench_with_input(BenchmarkId::new("native", size), &native_data, |b, data| {
             b.iter(|| {
@@ -135,20 +81,7 @@ fn bench_square(c: &mut Criterion) {
     let mut group = c.benchmark_group("goldilocks_square");
 
     for size in SIZES {
-        let mont_data = rand_mont_elements(size);
         let native_data = rand_native_elements(size);
-
-        group.bench_with_input(
-            BenchmarkId::new("montgomery", size),
-            &mont_data,
-            |b, data| {
-                b.iter(|| {
-                    for (x, _) in data {
-                        black_box(black_box(x).square());
-                    }
-                })
-            },
-        );
 
         group.bench_with_input(BenchmarkId::new("native", size), &native_data, |b, data| {
             b.iter(|| {
@@ -168,20 +101,7 @@ fn bench_inv(c: &mut Criterion) {
     const INV_SIZES: [usize; 4] = [100, 1000, 10000, 100000];
 
     for size in INV_SIZES {
-        let mont_data = rand_mont_elements(size);
         let native_data = rand_native_elements(size);
-
-        group.bench_with_input(
-            BenchmarkId::new("montgomery", size),
-            &mont_data,
-            |b, data| {
-                b.iter(|| {
-                    for (x, _) in data {
-                        black_box(black_box(x).inv().unwrap());
-                    }
-                })
-            },
-        );
 
         group.bench_with_input(BenchmarkId::new("native", size), &native_data, |b, data| {
             b.iter(|| {
