+++ crypto/math/src/field/fields/mod.rs
@@ -4,8 +4,6 @@ pub mod fft_friendly;
 pub mod montgomery_backed_prime_fields;
 /// Implementation of the Goldilocks Prime field (p = 2^448 - 2^224 - 1)
 pub mod u32_montgomery_backend_prime_field;
-/// Implementation of the u64 Goldilocks Prime field (p = 2^64 - 2^32 + 1)
-pub mod u64_goldilocks_field;
 /// Implementation of prime fields over 64 bit unsigned integers.
 pub mod u64_prime_field;
