+++ crypto/math/src/tests/mod.rs
@@ -18,7 +18,6 @@ pub mod msm_tests;
 pub mod polynomial_tests;
 pub mod short_weierstrass_point_tests;
 pub mod test_fields_tests;
-pub mod u64_goldilocks_field_tests;
 pub mod u64_prime_field_tests;
 pub mod unsigned_integer_montgomery_tests;
 pub mod unsigned_integer_u256_tests;
