+++ crypto/stark/benches/prover_benchmark.rs
@@ -2,7 +2,7 @@ use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
 use crypto::fiat_shamir::default_transcript::DefaultTranscript;
 use math::field::element::FieldElement;
 use math::field::fields::fft_friendly::extensions_goldilocks::Degree3GoldilocksExtensionField;
-use math::field::fields::fft_friendly::u64_goldilocks::U64GoldilocksPrimeField;
+use math::field::fields::fft_friendly::u64_goldilocks::GoldilocksField;
 use stark::examples::fibonacci_multi_column::{
     FibonacciMultiColumnAIR, FibonacciMultiColumnPublicInputs, compute_trace, create_public_inputs,
 };
@@ -11,7 +11,7 @@ use stark::proof::stark::StarkProof;
 use stark::prover::{IsStarkProver, Prover};
 use stark::verifier::{IsStarkVerifier, Verifier};
 
-type F = U64GoldilocksPrimeField;
+type F = GoldilocksField;
 type E = Degree3GoldilocksExtensionField;
 type FE = FieldElement<F>;
 
@@ -41,10 +41,10 @@ const QUICK_CONFIGS: &[BenchConfig] = &[
 
 /// Thorough benchmark configurations
 const THOROUGH_CONFIGS: &[BenchConfig] = &[
-    BenchConfig::new("fib_8col_16k", 8, 16384),
+    //BenchConfig::new("fib_8col_16k", 8, 16384),
     BenchConfig::new("fib_12col_16k", 12, 16384),
-    BenchConfig::new("fib_16col_8k", 16, 8192),
-    BenchConfig::new("fib_16col_16k", 16, 16384),
+    //BenchConfig::new("fib_16col_8k", 16, 8192),
+    //BenchConfig::new("fib_16col_16k", 16, 16384),
 ];
 
 /// Creates initial values for the specified number of columns
